+++ examples/libsql-n3-warm-container/hexastore.ts
@@ -1,8 +1,7 @@
 import { createClient } from "@libsql/client";
 import { QueryEngine } from "@comunica/query-sparql-rdfjs-lite";
-import { Client } from "@worlds/client";
-import { createComunicaLibsqlSparqlEngineFactory } from "@worlds/client/adapters/comunica";
-import { createLibsqlAdapter } from "@worlds/client/adapters/libsql";
+import type { Client } from "@worlds/client";
+import { createLibsqlComunicaClient } from "@worlds/client/adapters/libsql/comunica";
 import { DataFactory } from "n3";
 
 const { quad, namedNode, literal } = DataFactory;
@@ -17,14 +16,10 @@ const queryEngine = new QueryEngine();
 let warmIsolateClient: Client | undefined;
 
 async function getWarmIsolateClient(): Promise<Client> {
-  warmIsolateClient ??= new Client(
-    await createLibsqlAdapter({
-      client: databaseClient,
-      createSparqlEngine: createComunicaLibsqlSparqlEngineFactory({
-        queryEngine,
-      }),
-    }),
-  );
+  warmIsolateClient ??= await createLibsqlComunicaClient({
+    client: databaseClient,
+    queryEngine,
+  });
   return warmIsolateClient;
 }